Fix ROSBag Timezone issues
- Rosbag saves out bags with a timestamp in the name, this timestamp is
in the system timezone. When we were parsing and using this time Ros
time was assuming it was in UTC which was causing all bags to be deleted
as it thought they were >8 hours old. Add a test to reproduce this issue
and fix file_utils to convert the timestamps to UTC instead.