Delete bags integration test (#73)
* Periodic File Deleter Integration Test

- Integration test for the periodic file deleter that runs on a schedule
and cleans up rosbags older than a certain age.
- Make periodic deleter run correctly.
- Added regex for parsing timestamp out of rosbag names and test for a
bag that failed in integration testing
- Fix bug where bags were being deleted when the timestamp can't be parsed.

* Fix ROSBag Timezone issues

- Rosbag saves out bags with a timestamp in the name, this timestamp is
in the system timezone. When we were parsing and using this time Ros
time was assuming it was in UTC which was causing all bags to be deleted
as it thought they were >8 hours old. Add a test to reproduce this issue
and fix file_utils to convert the timestamps to UTC instead.

* Add passing bag rollover and max record time as parameters

- Add parameters to rolling recorder so that the rosbag rollover time
and rosbag max record time can be passed in via the launch file.
- Use these parameters in the test_periodic_file_deleter integration
test.

* Code style fixes

* Use RollingRecorderTestBase to minimize code duplication

- Rebase on another PR that created a RollingRecorderTestBase so this
can use it's methods instead of redefining them.

* Pass write_directory as parameter, fix test code

- Pass write_directory into rolling_recoder from the launch file instead
of hard coding it.
- Remove some logs added for testing and change some to DEBUG instead of
INFO.

* Add tests to GHA workflow, add timeout to wait function

- Add all rolling recorder tests to the Github Actions workflow so they actually run
- Add timeout to wait_for_rolling_recorder_nodes so it doesn't hang
forever in failure.

* Create specified rosbag recorder directory if it doesn't exist

* Remove TODO because clang-tidy doesn't like it

* Use default permissions for rosbag directory